+++ ui-test/ui-test.jsx
@@ -10,6 +10,7 @@ import {
     Checkbox,
     Error,
     Dialog,
+    ConfirmationDialog,
     BaseDialog,
     Icon,
     Info,
@@ -32,6 +33,7 @@ const Page = React.createClass({
         return {
             dialog: false,
             baseDialog: false,
+            confirmationDialog: false,
             timelineItems: [
                 {
                     id: 'http://example.com/1',
@@ -77,6 +79,13 @@ const Page = React.createClass({
     },
     tabClick(tabName) {
         console.log('tabClick:', tabName);
+    },
+    openConfirmationDialog() {
+        this.setState({confirmationDialog: true});
+    },
+    closeConfirmationDialog(param) {
+        console.log('ConfirmationDialog closed', param);
+        this.setState({confirmationDialog: false});
     },
     // template rendering
     render() {
@@ -161,6 +170,36 @@ const Page = React.createClass({
                 <p>Dialog Content</p>
                 <p>Dialog Content</p>
             </Dialog>
+        );
+
+        const testConfirmationDialog = (
+            <ConfirmationDialog title="ConfirmationDialog Title"
+                    active={this.state.confirmationDialog}
+                    modal={true}
+                    size="mini"
+                    cancelButton={<Button onClick={this.closeConfirmationDialog.bind(null, 'Cancel')}>Cancel</Button>}
+                    confirmButton={<Button onClick={this.closeConfirmationDialog.bind(null, 'Yes')}>Yes</Button>}
+            >
+                <p>ConfirmationDialog Content</p>
+                <p>ConfirmationDialog Content</p>
+                <p>ConfirmationDialog Content</p>
+                <p>ConfirmationDialog Content</p>
+                <p>ConfirmationDialog Content</p>
+                <p>ConfirmationDialog Content</p>
+                <p>ConfirmationDialog Content</p>
+                <p>ConfirmationDialog Content</p>
+                <p>ConfirmationDialog Content</p>
+                <p>ConfirmationDialog Content</p>
+                <p>ConfirmationDialog Content</p>
+                <p>ConfirmationDialog Content</p>
+                <p>ConfirmationDialog Content</p>
+                <p>ConfirmationDialog Content</p>
+                <p>ConfirmationDialog Content</p>
+                <p>ConfirmationDialog Content</p>
+                <p>ConfirmationDialog Content</p>
+                <p>ConfirmationDialog Content</p>
+                <p>ConfirmationDialog Content</p>
+            </ConfirmationDialog>
         );
 
         const testBaseDialog = (
@@ -283,7 +322,7 @@ const Page = React.createClass({
                     <h4 className="mdl-card__title-text">Test Buttons</h4>
                 </div>
                 <div className="mdl-card__content">
-                    {testDialog}{testBaseDialog}
+                    {testDialog}{testBaseDialog}{testConfirmationDialog}
                     <h5>Buttons using canonical icons</h5>
                     <Button colored iconName="edit" tooltip="own tooltip"/>
                     <Button accent iconName="delete" tooltip={false} />
@@ -298,11 +337,12 @@ const Page = React.createClass({
                     <Button iconName="filter" />
                     <Button iconName="sort" />
                     <Button iconName="hide" />
-                    <Button iconName="access_forbidden" />
+                    <Button iconName="access_forbidden" />ConfirmationDialog
                 </div>
                 <div className="mdl-card__actions">
                     <Button raised={true} accent ripple={false} onClick={this.openDialog}>Open Dialog</Button>
                     <Button raised={true} accent ripple={false} onClick={this.openBaseDialog}>Open BaseDialog</Button>
+                    <Button raised={true} accent ripple={false} onClick={this.openConfirmationDialog}>Open ConfirmationDialog</Button>
                     <Button raised={true} ripple={false} tooltip="This is a Test!" fabSize="mini">
                         <Icon name="mood" />
                     </Button>